Adds save button to local dumps view
