+++ frontend/src/Home.vue
@@ -39,8 +39,10 @@
 </div>
 </template>
 <script>
+import SavesMixin from "./SavesMixin";
 export default {
   name: "home",
+  mixins: [SavesMixin],
   data () {
     return {
       fields: [
@@ -107,32 +109,6 @@ export default {
         .catch(e => {
           this.notifyError(e);
         });
-    },
-    save: function (item) {
-      this.$set(item, "loading", true);
-      this.$http
-        .post("/api/save/" + item.pid + "/" + item.recording)
-        .then(response => {
-          item.loading = false;
-          this.$notify({
-            title: "Success",
-            text: "Dump saved to S3 storage: " + response.data.key,
-            type: "success"
-          });
-          console.log(response);
-        })
-        .catch(e => {
-          this.notifyError(e);
-          item.loading = false;
-        });
-    },
-    notifyError: function (error) {
-      this.$notify({
-        title: error.message,
-        text: error.response.data,
-        type: "error",
-        duration: 10000
-      });
     }
   },
   mounted () {
